Upgrade dependencies and modernize syntax:
- Updated project dependencies to the latest versions, including `Microsoft.Extensions` and `NSubstitute`.
- Added `Microsoft.SourceLink.GitHub` for better debugging support.
- Refactored to adopt modern C# syntax with target-typed new expressions and inline static usage for improved clarity and efficiency.

+++ src/SolSharp.Programs/UpgradeableBpfLoaderProgram.cs
@@ -45,7 +45,7 @@ public static class UpgradeableBpfLoaderProgram
         ulong lamports,
         ulong programLength)
     {
-        var space = checked((ulong)UpgradeableBpfLoaderState.BufferMetadataLength + programLength);
+        var space = checked(UpgradeableBpfLoaderState.BufferMetadataLength + programLength);
         return
         [
             SystemProgram.CreateAccount(payer, buffer, lamports, space, ProgramId),
